内容:GLFloat 写成了CGFloat 导致在iPhone 6+机器上crash

+++ framework/YDGLOperationKit/source/core/YDGLOperationNodeView.m
@@ -281,7 +281,7 @@
         }; break;
     }
     
-    CGFloat *imageVertices=malloc(12*sizeof(GLfloat));
+    GLfloat *imageVertices=malloc(12*sizeof(GLfloat));
     
     imageVertices[0] = -widthScaling;
     imageVertices[1] = -heightScaling;
